API reference clean up
