+++ src/views/UserProfileView.vue
@@ -4,7 +4,7 @@ import { useRoute, useRouter } from 'vue-router'
 import { NButton, NText, NIcon } from 'naive-ui'
 import { MdTrash, MdLogOut } from '@vicons/ionicons4'
 import { UserFollow } from '@vicons/carbon'
-import { API } from '@/constants'
+import API from '@/constants/api'
 import { myId, isLoggedIn } from '@/constants/authState'
 import SearchBar from '@/components/SearchBar.vue'
 import FileList from '@/components/FileList.vue'
